requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
pymysql
python-dotenv
Pillow
gunicorn
Flask-JWT-Extended
numpy
tensorflow-cpu